backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgi-correlation-id==4.3.1
boto3==1.34.92
botocore==1.34.92
click==8.1.7
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
jmespath==1.0.1
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.9
PyYAML==6.0.1
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
